fix(server): ssl certs fix & listener api
now the right ssl certs are going to be used where previously there was
a bug that, even after specificying the custom ssl cert, a new ssl cert
is going to be generated and used.
Listener api has been updated and worked on to add a way to stop
listeners registered via the plugin api.